feat(dashboard-workspaces): replace overlay concept with singleton dashboard workspaces
Settings and Claude Settings now spawn as draggable singleton workspace rows
instead of floating overlays. Extension API gains registerDashboardWorkspace
and drops registerOverlay. WorkspaceItem renders a per-registry icon; TitleBar
places extension-contributed buttons left of the gear. MCP list_overlays tool
replaced by list_dashboard_workspaces. All overlay infrastructure deleted.
